requirements.txt
# GUI Framework
PyQt6==6.6.1
PyQt6-Charts==6.6.0
PyQt6-WebEngine==6.6.0
PyQt6-Multimedia==6.6.0

# Database
SQLAlchemy==2.0.23
alembic==1.13.0

# AI & ML
openai==1.3.0
anthropic==0.8.0
langchain==0.0.350
langchain-community==0.0.38

# Computer Vision
opencv-python==4.8.1
ultralytics==8.0.220
Pillow==10.1.0
numpy==1.26.2

# Smartwatch Integration
requests==2.31.0
requests-oauthlib==1.3.1
websockets==12.0
aiohttp==3.9.1

# Data & Visualization
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
cryptography==41.0.7
APScheduler==3.10.4
loguru==0.7.2
pydantic==2.5.0

# Development
pytest==7.4.3
pytest-qt==4.2.0
black==23.12.0
pylint==3.0.3
